+++ crates/core/src/tools_config.rs
@@ -1,13 +1,16 @@
-use crate::errors::ProtoError;
+use crate::{errors::ProtoError, plugin::PluginLocator};
+use convert_case::{Case, Casing};
 use rustc_hash::FxHashMap;
 use starbase_utils::toml::{self, TomlTable, TomlValue};
 use std::path::{Path, PathBuf};
+use std::str::FromStr;
 
 pub const TOOLS_CONFIG_NAME: &str = ".prototools";
 
 #[derive(Debug, Default)]
 pub struct ToolsConfig {
     pub tools: FxHashMap<String, String>,
+    pub plugins: FxHashMap<String, PluginLocator>,
     pub path: PathBuf,
 }
 
@@ -46,27 +49,51 @@ impl ToolsConfig {
 
         let config: TomlValue = toml::read_file(path)?;
         let mut tools = FxHashMap::default();
+        let mut plugins = FxHashMap::default();
 
         if let TomlValue::Table(table) = config {
             for (key, value) in table {
-                if let TomlValue::String(version) = value {
-                    tools.insert(key, version);
-                } else {
-                    return Err(ProtoError::InvalidConfig(
-                        path.to_path_buf(),
-                        format!("Expected a version string for \"{key}\"."),
-                    ));
+                match value {
+                    TomlValue::String(version) => {
+                        tools.insert(key, version);
+                    }
+                    TomlValue::Table(plugins_table) => {
+                        for (plugin, locator) in plugins_table {
+                            if let TomlValue::String(locator) = locator {
+                                plugins.insert(
+                                    plugin.to_case(Case::Kebab),
+                                    PluginLocator::from_str(&locator)?,
+                                );
+                            } else {
+                                return Err(ProtoError::InvalidConfig(
+                                    path.to_path_buf(),
+                                    format!(
+                                        "Invalid plugin \"{plugin}\", expected a locator string."
+                                    ),
+                                ));
+                            }
+                        }
+                    }
+                    _ => {
+                        return Err(ProtoError::InvalidConfig(
+                            path.to_path_buf(),
+                            format!(
+                                "Invalid field \"{key}\", expected a mapped tool version, or a [plugins] map."
+                            ),
+                        ))
+                    }
                 }
             }
         } else {
             return Err(ProtoError::InvalidConfig(
                 path.to_path_buf(),
-                "Expected a mapping of tools to versions.".into(),
+                "Expected a mapping of tools or plugins.".into(),
             ));
         }
 
         Ok(ToolsConfig {
             tools,
+            plugins,
             path: path.to_owned(),
         })
     }
@@ -77,6 +104,16 @@ impl ToolsConfig {
 
         for (tool, version) in &self.tools {
             map.insert(tool.to_owned(), TomlValue::String(version.to_owned()));
+        }
+
+        if !self.plugins.is_empty() {
+            let mut plugins = TomlTable::with_capacity(self.plugins.len());
+
+            for (plugin, locator) in &self.plugins {
+                plugins.insert(plugin.to_owned(), TomlValue::String(locator.to_string()));
+            }
+
+            map.insert("plugins".to_owned(), TomlValue::Table(plugins));
         }
 
         toml::write_file(&self.path, &TomlValue::Table(map), true)?;